code/requirements.txt
absl-py
aiohttp
aiosignal
alembic==1.12.0
anyio
appdirs==1.4.4
argon2-cffi
argon2-cffi-bindings
asttokens
async-timeout
attrs
Babel
backcall
beautifulsoup4
bleach
blinker==1.4
Bottleneck
brotlipy==0.7.0
cachetools
certifi
cffi
charset-normalizer
click
cloudpickle
cmaes==0.10.0
colorama
colorlog==6.7.0
comm
contourpy==1.1.0
cryptography
cycler==0.11.0
cytoolz
dask
debugpy
decorator
defusedxml
docker-pycreds==0.4.0
easydict==1.9
einops==0.6.1
entrypoints
executing
Farama-Notifications
fastjsonschema
filelock
fonttools==4.40.0
frozenlist
fsspec
gitdb==4.0.10
GitPython==3.1.32
gmpy2
google-auth
google-auth-oauthlib
greenlet==2.0.2
grpcio
gymnasium
h5py
huggingface-hub==0.16.4
idna
imagecodecs
imageio
importlib-metadata
importlib-resources==6.0.0
ipykernel
ipython
ipython-genutils
jax-jumpy
jedi
Jinja2
joblib
json5
jsonschema
jupyter-server
jupyter-server-mathjax==0.2.6
jupyter_client
jupyter_core
jupyterlab
jupyterlab-pygments
jupyterlab_server
kiwisolver==1.4.4
llvmlite
locket
Mako==1.2.4
Markdown
MarkupSafe
matplotlib==3.7.2
matplotlib-inline
mistune
mkl-fft==1.3.1
mkl-random
mkl-service==2.4.0
mpmath==1.2.1
multidict
nbclassic==1.0.0
nbclient
nbconvert
nbdime==3.2.1
nbformat
nest-asyncio
networkx
notebook
notebook_shim==0.2.3
numba
numexpr
numpy
oauthlib
opencv-python==4.8.0.74
optuna==3.3.0
packaging
pandas==1.5.3
pandocfilters
parso
partd
pathtools==0.1.2
pexpect
pickleshare
Pillow==9.4.0
platformdirs
prometheus-client
prompt-toolkit
protobuf==3.20.3
psutil
ptyprocess
pure-eval
pyasn1
pyasn1-modules==0.2.8
pycparser
Pygments
PyJWT
pyOpenSSL
pyparsing==3.0.9
pyrsistent
PySocks
python-dateutil
pytz
PyWavelets
PyYAML
pyzmq
regex==2023.6.3
requests
requests-oauthlib==1.3.0
rsa
safetensors==0.3.1
scikit-image
scikit-learn
scipy
seaborn==0.12.2
Send2Trash
sentry-sdk==1.28.0
setproctitle==1.3.2
six
smmap==5.0.0
sniffio
soupsieve
SQLAlchemy==2.0.21
stack-data
sympy
tensorboard
tensorboard-data-server
tensorboard-plugin-wit
tensorboardX==2.6.2.2
termcolor==2.3.0
terminado
threadpoolctl
tianshou
tifffile
tinycss2
tokenizers==0.13.3
tomli
toolz
torch==2.0.0
torchaudio==2.0.0
torchvision==0.15.0
tornado
tqdm
traitlets
transformers==4.30.2
triton==2.0.0
typing_extensions
urllib3
wandb==0.15.5
wcwidth
webencodings==0.5.1
websocket-client
Werkzeug
yarl
zipp